Refactored test cases into test_data_collection.py.
